+++ CHANGELOG.md
@@ -37,6 +37,8 @@ end
     * `paddle_pay ~> 0.2`
     * `receipts ~> 2`
 
+* Stripe.max_network_retries is now set to 2 by default. - @excid3
+  This adds idempotency keys automatically to each request so that they can be safely retried.
 * Stripe Subscriptons can now be paused and resumed - @excid3
 * Replace `update_email!` with `update_customer!` - @excid3
 * Add options for `cancel_now!` to support `invoice_now` and `prorate` flags for Stripe - @excid3